requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
gunicorn==22.0.0
Werkzeug==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
requests==2.32.3
Click==8.1.7
Flask-Login==0.6.3
Flask-Migrate==4.1.0
pg8000==1.31.5
redis==7.1.0
Flask_Caching==2.3.1
gevent==25.9.1



